Use concrete type for struct

+++ src/types/customer_type.jl
@@ -5,7 +5,7 @@ export Customer
 export id
 
 # code
-struct Customer{T <: Union{String, Int}}
+struct Customer{T <: Union{String, Int64}}
 
     customer_id::T
 
@@ -13,7 +13,7 @@ end
 
 
 """
-    function Base.convert(::Type{Customer}, customer_id::T) where {T <: Union{AbstractString, Int}}
+    function Base.convert(::Type{Customer}, customer_id::T) where {T <: Union{AbstractString, Int64}}
 
 Single-line funcition to automatically convert permissible types to Customer.
 Overload function in Base module.
@@ -22,7 +22,7 @@ first argument:     Customer (user defined type)
 customer_id:        ID to be converted to Customer type
 """
 
-Base.convert(::Type{Customer}, customer_id::T) where {T <: Union{AbstractString, Int}} = Customer(customer_id)
+Base.convert(::Type{Customer}, customer_id::T) where {T <: Union{AbstractString, Int64}} = Customer(customer_id)
 
 """
     function id(obj::Customer)
